+++ app/containers/ContactEdit/messages.js
@@ -1,0 +1,25 @@
+/*
+ * ContactEdit Messages
+ *
+ * This contains all the text for the ContactEdit component.
+ */
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  saveBtn: {
+    id: 'app.containers.ContactEdit.save',
+    defaultMessage: 'Save Changes',
+  },
+  createBtn: {
+    id: 'app.containers.ContactEdit.create',
+    defaultMessage: 'Create',
+  },
+  cancelBtn: {
+    id: 'app.containers.ContactEdit.cancel',
+    defaultMessage: 'Cancel',
+  },
+  abandonChanges: {
+    id: 'app.containers.ContactEdit.abandonChanges',
+    defaultMessage: 'Abandon Changes?',
+  },
+});